updated check for validate_item_esd_ddlm
Previously, if a dataitem was a Measurand and did not have an SU, then it failed the test. Given the Error text for "validate_item_esd" being "Error: The following data items should not have esds appended", I reasoned that the similar condition should apply for `validate_item_esd_ddlm`.

The error is now triggered if an SU is given to a datavalue that is not a measurand. It is not triggered if an SU is not given to a value that is a measurand.